Refactor desktop and extension apps to use CoreApp component; remove AppProviders. Update core state management and integrate vault creation logic into the new structure.
